==> models/edge-3-day-hosp-v1/src/notebooks/00_get_raw_data.ipynb <==
## SET the CLIENT once in shared/constants/CLIENT before executing 

import sys
sys.path.insert(0, '/src')
from shared.load_raw_data import fetch_training_data, load_raw_data_from_files
from shared.utils import get_client_class
from shared.constants import CLIENT
from eliot import to_file
to_file(sys.stdout)

%load_ext autoreload
%autoreload 2

import os

# clientClass = get_client_class(client=CLIENT)
# TRAIN_START_DATE, TRAIN_END_DATE = getattr(clientClass(), 'get_training_dates')()

# print(CLIENT)
# print(os.environ.get('SAIVA_ENV','dev'))
# print(TRAIN_START_DATE, TRAIN_END_DATE)

TRAIN_START_DATE = '2019-01-01'
TRAIN_END_DATE = '2020-02-28'

### =============== Delete all files inside /data/processed & /data/raw ==================

import shutil
from pathlib import Path

# processed_path = Path('/data/processed')
# raw_path = Path('/data/raw')
# shutil.rmtree(processed_path)
# shutil.rmtree(raw_path)
# print("Success.......")

### ======================== Fetch Data ============================

# Loads the data from SQL db and store them in local directory as cache
# Can pass multiple clients of single

result_dict = fetch_training_data(
    client_list=['infinity-benchmark', 'infinity-infinity'],
    train_start_date=TRAIN_START_DATE,
    train_end_date=TRAIN_END_DATE
)

# print(result_dict['master_patient_lookup'].shape)
print(result_dict['patient_census'].shape)
print(result_dict['patient_rehosps'].shape)
print(result_dict['patient_demographics'].shape)
print(result_dict['patient_diagnosis'].shape)
print(result_dict['patient_vitals'].shape)
print(result_dict['patient_meds'].shape)
print(result_dict['patient_orders'].shape)
print(result_dict['patient_alerts'].shape)
print(result_dict['patient_progress_notes'].shape)
print(result_dict['stays'].shape)
print(result_dict.keys())

# have a max of 15042 master_patient_lookup rows ie. Infinity-Infinity

### ======================== TESTING ==========================

# Once fetch_training_data loads the data, use the same cache 

data_dict = load_raw_data_from_files(dirname='/data/raw')
# print(data_dict['master_patient_lookup'].shape)
print(data_dict['patient_census'].shape)
print(data_dict['patient_rehosps'].shape)
print(data_dict['patient_demographics'].shape)
print(data_dict['patient_diagnoses'].shape)
print(data_dict['patient_vitals'].shape)
print(data_dict['patient_meds'].shape)
print(data_dict['patient_orders'].shape)
print(data_dict['patient_alerts'].shape)
print(data_dict['patient_progress_notes'].shape)
print(data_dict['stays'].shape)
print(data_dict.keys())

# TESTING specific queries

# query=f"""
#         select distinct patientid, facilityid, orderdate, gpiclass, 
#         gpisubclassdescription, orderdescription, pharmacymedicationname, a.PhysiciansOrderID
#         from view_ods_physician_order_list_v2 a
#         inner join view_ods_physician_order_list_med b
#         on a.PhysiciansOrderID = b.PhysiciansOrderID 
#         where orderdate between '{train_start_date}' and '{test_end_date}';
#         """

# df = pd.read_sql(query, con=client_sql_engine)
# print(df.shape)
# df.head()
