==> models/edge-3-day-hosp-v1/src/notebooks/02_more_featurizers.ipynb <==
# More features... 
* Demographics (age, race, etc)
* Previous stays
* Notes

import sys
import pickle as pkl
import scipy

%load_ext autoreload
%autoreload 2

sys.path.append('/src')
from shared.featurizers import Featurizer2
from shared.load_raw_data import load_raw_data_from_files
from eliot import to_file
to_file(sys.stdout)


data_dict = load_raw_data_from_files('/data/raw')

with open("/data/processed/prediction_times.pkl", 'rb') as f_in: 
    prediction_times = pkl.load(f_in)

    
with open('/data/processed/feature_colnames.pkl', 'rb') as f_in: feature1_names = pkl.load(f_in)
feature1_csr = scipy.sparse.load_npz('/data/processed/features_csr.npz')

%%time

featurizer2 = Featurizer2(data_dict, prediction_times)
feature_csr, feature_names = featurizer2.process(feature1_csr, feature1_names)

# Save for later use
scipy.sparse.save_npz('/data/processed/final_csr.npz', feature_csr)
with open('/data/processed/final_colnames.pkl', 'wb') as f_out:
    pkl.dump(feature_names, file=f_out)

## ========================END=======================
